RAG/requirements.txt
PyPDF2
flask
python-dotenv
langchain
langchain-openai
langchain-community
langchain-ollama
langchain-core
faiss-cpu
pytest
